Use offset and float32 to reduce coordinate memory footprint

+++ internal/data/datapoint.go
@@ -3,9 +3,9 @@ package data
 // Contains data of a Point Cloud Point, namely X,Y,Z coords,
 // R,G,B color components, Intensity and Classification
 type Point struct {
-	X              float64
-	Y              float64
-	Z              float64
+	X              float32
+	Y              float32
+	Z              float32
 	R              uint8
 	G              uint8
 	B              uint8
@@ -14,7 +14,7 @@ type Point struct {
 }
 
 // Builds a new Point from the given coordinates, colors, intensity and classification values
-func NewPoint(X, Y, Z float64, R, G, B, Intensity, Classification uint8) *Point {
+func NewPoint(X, Y, Z float32, R, G, B, Intensity, Classification uint8) *Point {
 	return &Point{
 		X:              X,
 		Y:              Y,
